Fix doc/code mismatches and use c_size_t for refcount reads
Update design doc and tasks.json to reflect actual implementation:
three bridge intrinsics with Python-only _read_refcount() helper
instead of four intrinsics with refcount_of_meminfo_intp. Change
_read_refcount to use ctypes.c_size_t matching the MemInfo layout.

+++ test/test_ducklib.py
@@ -2925,7 +2925,7 @@ def _read_refcount(meminfo_intp):
     MemInfo.refct is the first field (size_t) — stable since numba 0.50.
     See numba/core/runtime/nrt.cpp.
     """
-    return ctypes.c_int64.from_address(meminfo_intp).value
+    return int(ctypes.c_size_t.from_address(meminfo_intp).value)
 
 
 # ---- Heap-owning structref (for nested-dtor test) ----
